+++ server_springboot/src/main/java/com/barataribeiro/taskr/repositories/relations/ProjectUserRepository.java
@@ -11,14 +11,14 @@ import java.util.Set;
 public interface ProjectUserRepository extends JpaRepository<ProjectUser, Long> {
     @EntityGraph(attributePaths = {"user"})
     @Query("select p from ProjectUser p where p.project.id = :id order by p.user.username")
-    Set<ProjectUser> findAllByProject_Id(Integer id);
+    Set<ProjectUser> findAllByProject_Id(Long id);
 
     @Query("""
             select (count(p) > 0) from ProjectUser p
             where p.project.id = :id and p.user.id = :userId and p.isProjectManager = :isProjectManager""")
-    boolean existsProjectWhereUserByIdIsManager(Integer id, String userId, boolean isProjectManager);
+    boolean existsProjectWhereUserByIdIsManager(Long id, String userId, boolean isProjectManager);
 
     void deleteByProject(Project project);
 
-    boolean existsByProject_IdAndUser_Id(Integer id, String userId);
+    boolean existsByProject_IdAndUser_Id(Long id, String userId);
 }
